requirements.txt
selenium
pillow
flask
flask-cors
opencv-python
numpy
pyinstaller
requests 